backend/qwen_api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
qdrant-client==1.7.0
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.0
groq==0.4.1